requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
black==21.6b0
boto3==1.17.53
botocore==1.20.53
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==8.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-ses==2.0.0
django-storages==1.11.1
django-taggit==1.4.0
django-taggit-serializer==0.1.7
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
future==0.18.2
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
msgpack==1.0.2
mux-python==1.12.0
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.7
sentry-sdk==1.1.0
service-identity==21.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stripe==2.57.0
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
zope.interface==5.4.0
